Remove another detection

+++ wechat.sh
@@ -6,10 +6,7 @@ if [ -f /usr/lib/wechat/alt.desktop ]; then
 		~/.local/share/applications/com.qq.weixin.desktop
 fi
 
-if [ -f /usr/bin/portable ] && [ -f /usr/lib/wechat/portable-config ]; then
-	export _portalConfig=/usr/lib/wechat/portable-config
-	portable $@
-elif [ -f /usr/bin/portable ]; then # Keep this for users prior to the merge
+if [ -f /usr/bin/portable ] && [ ! -f /usr/lib/wechat/portable-config ]; then # Keep this for users prior to the merge
 	echo "You'll need to install wechat-bwrap for sandboxing to work!"
 	notify-send "You'll need to install wechat-bwrap for sandboxing to work!"
 	zenity --error --title "Sandbox non-functional" --icon=security-low-symbolic --text "Please install aur/wechat-bwrap"
